pyproject.toml
[tool.poetry]
name = "clicks-and-cookies"
version = "0.1.0"
description = "django link in bio app with clicks and cookies"
authors = ["flapjackstan <elm.s.camargo@gmail.com>"]
license = "MIT License"
readme = "README.md"
packages = [{include = "clicks_and_cookies"}]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.1.5"
python-dotenv = "^0.21.1"
djangorestframework = "^3.14.0"
django-cors-headers = "^3.13.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
